+++ CHANGELOG.md
@@ -22,6 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add release version on the interface (dashboard and page footer) #116
 - Editors' choice at time of submission: informative text added. (gitlab #369)
 - Browse by Volume or Sections : Handle new content type 'application/json' to return json instead of html
+- Fix #143 Allow to switch off manual or automatic DOI assignation (disabled mode)
+- [#37](https://github.com/CCSDForge/episciences/issues/37)
+  - On URLs like /volume/view/id/314159 and /section/view/id/314159 Handle new content type 'application/json' to return json instead of html (only published content)
+  - On URLs like /volume/edit?id=314159 Handle new content type 'application/json' to return json instead of html but including all statuses of articles ; an authenticated and allowed user is required
+  - JSON added to the list of public export formats
 - added: info message for the stats.
 
 ### Changed
@@ -41,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: [#117](https://github.com/CCSDForge/episciences/issues/117): An empty e-mail is not sent or received
 - Fixed [#129](https://github.com/CCSDForge/episciences/issues/129): when you do a search from the mail history page nothing happens.
 - Fixed [#141](https://github.com/CCSDForge/episciences/issues/141): visibility of pages menu is wrong.
+
 ### Changed
 - [#126](https://github.com/CCSDForge/episciences/issues/126): prevent modal closure: disabling the click outside the modal area and by pressing Esc
 - Enabling the "COI" option overrides the "Encapsulate editors" option